Implement #40 - apply sealed interface to stateMachine and organize packages of components
