+++ app/src/main/java/com/nlab/reminder/core/state/util/StateMachineConfig.kt
@@ -14,19 +14,11 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.state.impl
-
-import com.nlab.reminder.core.state.Action
-import com.nlab.reminder.core.state.ActionProcessor
-import com.nlab.reminder.core.state.State
-import com.nlab.reminder.core.state.StateMachine
-import kotlinx.coroutines.flow.*
+package com.nlab.reminder.core.state.util
 
 /**
  * @author Doohyun
  */
-internal class DefaultStateMachine<A : Action, S : State>(
-    actionProcessor: ActionProcessor<A>,
-    override val state: StateFlow<S>,
-) : StateMachine<A, S>,
-    ActionProcessor<A> by actionProcessor
+object StateMachineConfig {
+    var defaultExceptionHandler: (Throwable) -> Unit = DefaultExceptionHandler()
+}
